Masowa zmiana flag dla dokumentów

+++ engine/ajax/a_report_set_report_flags.php
@@ -16,7 +16,7 @@ class Ajax_report_set_report_flags extends CPage {
 	}*/
 	
 	function execute(){
-		global $mdb2, $user;
+		global $db, $mdb2, $user;
 
 		if (!intval($user['user_id'])){
 			throw new Exception("Brak identyfikatora użytkownika");
@@ -24,16 +24,37 @@ class Ajax_report_set_report_flags extends CPage {
 
 		$report_id = intval($_POST['report_id']);
 		$cflag_id = intval($_POST['cflag_id']);
-		$flag_id = intval($_POST['flag_id']);
-		
-		if ($flag_id){		
-			$sql = "REPLACE reports_flags SET corpora_flag_id={$cflag_id}, report_id={$report_id}, flag_id={$flag_id}";
-			$result = db_execute($sql);
-		}
-		else {
-			$sql = "DELETE FROM reports_flags WHERE corpora_flag_id={$cflag_id} AND report_id={$report_id}";
-			$result = db_execute($sql);
-		}
+                $flag_id = intval($_POST['flag_id']);
+                
+                $params = array('corpora_flag_id' => $cflag_id, 
+                                'report_id'       => $report_id,
+                                'flag_id'         =>$flag_id);
+               
+                
+                //Masowa zmiana statusu flagi 
+                if(isset($_POST['multiple']) === true){
+                    
+                    $document_ids = ($_POST['documents_ids']);
+                    if(empty($document_ids)){
+                        return;
+                    }
+                    
+                    foreach($document_ids as $document){
+                        $params['report_id'] = $document;
+                        $db->replace("reports_flags", $params);
+                    }
+             
+                }
+                //Zmiana jednej flagi
+                else {
+                    if ($flag_id){		
+                        $db->replace("reports_flags", $params);
+                    }
+                    else {
+                            $sql = "DELETE FROM reports_flags WHERE corpora_flag_id={$cflag_id} AND report_id={$report_id}";
+                            $result = db_execute($sql);
+                    }
+                }
 
 		return;
 	}
